requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
python-dotenv
openai>=1.0.0
numpy
python-docx
pypdf
python-multipart